+++ test/Get-Column.Tests.ps1
@@ -1,34 +1,34 @@
 #Requires -Modules @{ ModuleName='Pester'; ModuleVersion='5.0.0' }
 
-Describe 'Get-DacColumn' {
+param (
+    [System.IO.FileInfo] $DacPacFile = "$PsScriptRoot\sql-server-samples\samples\databases\wide-world-importers\wwi-ssdt\wwi-ssdt\bin\Debug\WideWorldImporters.dacpac"
+)
 
-    BeforeDiscovery {
-        [System.IO.FileInfo] $Global:DacPacFile = "$PsScriptRoot\sql-server-samples\samples\databases\wide-world-importers\wwi-ssdt\wwi-ssdt\bin\Debug\WideWorldImporters.dacpac"
-    }
+Describe Get-Column {
 
     BeforeAll {
         Import-Module $PSScriptRoot\..\publish\PsDac\PsDac.psd1 -ErrorAction Stop
     }
 
-    Context 'DacPac' -Skip:( -Not $Global:DacPacFile.Exists ) {
-        Context 'Model' {
+    Context DacPac -Skip:( -Not $DacPacFile.Exists ) {
+        Context Model {
             BeforeAll {
-                $Script:Model = Import-DacModel -Path $Global:DacPacFile
+                $Model = Import-DacModel -Path $DacPacFile
             }
             Context 'Table' {
                 BeforeAll {
-                    $Script:Table = $Script:Model | Get-DacTable -Name '[Application].[Cities]'
+                    $Table = $Model | Get-DacTable -Name '[Application].[Cities]'
                 }
 
                 It 'Returns all columns of a table' {
-                    $columns = $Script:Table | Get-DacColumn
-                    $columns | Should -Not -BeNullOrEmpty
-                    $columns.Count | Should -BeGreaterThan 3
+                    $Columns = $Table | Get-DacColumn
+                    $Columns | Should -Not -BeNullOrEmpty
+                    $Columns.Count | Should -BeGreaterThan 3
                 }
 
                 It 'Returns a column of a table by name' {
-                    $column = $Script:Table | Get-DacColumn -Name '[Application].[Cities].[CityID]'
-                    $column | Should -Not -BeNullOrEmpty
+                    $Column = $Table | Get-DacColumn -Name '[Application].[Cities].[CityID]'
+                    $Column | Should -Not -BeNullOrEmpty
                 }
             }
         }